Split old catch-all bot commands dir to separate dirs
