+++ bot-commands/legacy-commands/code-legacy.js
