+++ bot-commands/legacy-commands/os.test.js
