+++ bot-commands/legacy-commands/top-legacy.js
