+++ bot-commands/legacy-commands/top-legacy.test.js
