+++ bot-commands/legacy-commands/xy-legacy.test.js
