+++ bot-commands/points/points.js
@@ -1,7 +1,7 @@
 const axios = require('axios');
 const config = require('../../config');
 const { registerBotCommand } = require('../../botEngine');
-const club40Gifs = require('./club_40_gifs.json');
+const club40Gifs = require('../club-40/club_40_gifs.json');
 const { isAdmin } = require('../../utils/is-admin');
 
 axios.default.defaults.headers.common.Authorization = `Token ${config.pointsbot.token}`;
