+++ bot-commands/points/points.test.js
